features

+++ src/main/java/com/coinpurse/web/services/impl/RoleServicesImpl.java
@@ -1,5 +1,6 @@
 package com.coinpurse.web.services.impl;
 
+import com.coinpurse.web.domain.exceptions.ResourceNotFoundException;
 import com.coinpurse.web.infrastructure.client.CurrencyApiClient;
 import com.coinpurse.web.model.Role;
 import com.coinpurse.web.repository.RoleRepository;
@@ -9,6 +10,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+import static com.coinpurse.web.constants.ErrorMessages.ROLE_NOT_FOUND;
 
 @Service
 public class RoleServicesImpl implements RoleServices {
@@ -22,13 +27,31 @@ public class RoleServicesImpl implements RoleServices {
         this.roleRepository = roleRepository;
     }
 
+    @Override
     @Transactional
     public Role saveRole(Role role) {
         return roleRepository.save(role);
     }
 
+    @Override
     @Transactional
     public void deleteRole(Role role) {
-        roleRepository.deleteById(role.getId());
+        if(roleRepository.existsById(role.getId())) {
+            roleRepository.deleteById(role.getId());
+        } else {
+            throw new ResourceNotFoundException(ROLE_NOT_FOUND);
+        }
+    }
+
+    @Override
+    public List<Role> getAll() {
+        return roleRepository.findAll();
+    }
+
+    @Override
+    public Role getRoleById(Long roleId) {
+        return roleRepository.findById(roleId).orElseThrow(
+                () -> new ResourceNotFoundException(ROLE_NOT_FOUND)
+        );
     }
 }
